need to program fall Rect

+++ WindowProgramming/CFramework.cpp
@@ -60,7 +60,7 @@ void CFramework::Update()
 	++m_iFrame;
 	if (m_fTime - m_fBaseTime > 1000) {
 		cout << "FPS : " << m_iFrame << endl;
-		m_iFrame = m_iFrame * 1000.0 / (m_fTime - m_fBaseTime);
+		m_iFrame = static_cast<int>(m_iFrame * 1000.0f / (m_fTime - m_fBaseTime));
 		m_fBaseTime = m_fTime;
 		m_iFrame = 0;
 	}
